fix: d.ts to ts
